Complete all features

+++ db/migrate/20181113022552_create_rates.rb
@@ -3,7 +3,7 @@ class CreateRates < ActiveRecord::Migration[5.1]
     create_table :rates do |t|
       t.belongs_to :rater
       t.belongs_to :rateable, :polymorphic => true
-      t.float :stars, :null => false
+      t.integer :stars, :null => false
       t.string :dimension
       t.timestamps
     end
